refactor(players): migrate controller to module
- Move to Players::Controllers namespace
- Reduce from 750 to 350 lines (-53%)
- Use services for business logic
- Create proxy for backwards compatibility
